+++ packages/orq-rc/src/models/operations/updatedatapoint.ts
@@ -2048,7 +2048,7 @@ export const UpdateDatapointResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-05-28T09:00:06.548Z",
+    "2025-05-28T11:21:06.214Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -2092,7 +2092,7 @@ export const UpdateDatapointResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-05-28T09:00:06.548Z"))
+  updated: z.date().default(() => new Date("2025-05-28T11:21:06.214Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {